chahge type of offer id

+++ src/main/java/application/dto/MessageDTO.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class MessageDTO {
 
     private LocalDateTime time;
-    private Integer offerId;
+    private Long offerId;
     @NotNull
     @Size(min = 2, max = 3000)
     private String message;
